feat: migrate user profile form to a modal for consistency
