+++ src/components/index.ts
@@ -4,7 +4,6 @@ export * from "./formats";
 export * from "./modals";
 export * from "./pages";
 export * from "./popovers";
-export * from "./profile";
 export * from "./selectors";
 export * from "./tables";
 export * from "./toolbar";
