+++ src/components/modals/index.ts
@@ -1,3 +1,4 @@
 export * from "./DeckModal";
 export * from "./FormatModal";
 export * from "./MatchModal";
+export * from "./UserProfileModal"
