--- src/components/profile/index.ts
@@ -1,2 +1,0 @@
-export * from "./UserDecksCard";
-export * from "./UserProfileForm";
